+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * **Breaking**: the `--workspace` flag is no longer set for all cargo commands
   by default. The previous behavior can be recovered by setting `cargoExtraArgs
   = "--workspace";` in any derivation.
+* All cargo invocations made during the build are automatically logged
 
 ## [0.5.1] - 2022-07-20